+++ src/components/Listbox/SelectStyled.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import Select from '.'
+import Select from './Listbox'
 import { IconBook } from '../../index'
 
 export default {
@@ -34,26 +34,6 @@ export const Default = (args: any) => (
     })}
   </Select>
 )
-
-// export const ErrorState = (args: any) => (
-//   <Select
-//     label="Choose a person"
-//     descriptionText="Choose a person for this role"
-//   >
-//     {people.map((person) => {
-//       return (
-//         <Select.Option
-//           value={person.value}
-//           addOnBefore={({ active, selected }: any) => [
-//             <img src={person.avatar} alt="" className="h-6 w-6 rounded-full" />,
-//           ]}
-//         >
-//           {person.label}
-//         </Select.Option>
-//       )
-//     })}
-//   </Select>
-// )
 
 const people = [
   {
@@ -187,159 +167,10 @@ export const ErrorState = (args: any) => (
   </Select>
 )
 
-// export const withOptionGroup = (args: any) => (
-//   <Select {...args}>
-//     <OptGroup label="languages">
-//       <Option value="javascript">JavaScript</Option>
-//       <Option value="typeScript">TypeScript</Option>
-//     </OptGroup>
-//     <OptGroup label="libaries">
-//       <Option value="react">React</Option>
-//     </OptGroup>
-//   </Select>
-// )
-
-// export const withIcon = (args: any) => (
-//   <Select {...args}>
-//     <Option value="javascript">JavaScript</Option>
-//     <Option value="typeScript">TypeScript</Option>
-//     <Option value="react">React</Option>
-//   </Select>
-// )
-
-// export const withOptionLabel = (args: any) => (
-//   <Select {...args}>
-//     <Option value="javascript">JavaScript</Option>
-//     <Option value="typeScript">TypeScript</Option>
-//     <Option value="react">React</Option>
-//   </Select>
-// )
-
-// export const withBeforeAndAfterLabel = (args: any) => (
-//   <Select {...args}>
-//     <Option value="javascript">JavaScript</Option>
-//     <Option value="typeScript">TypeScript</Option>
-//     <Option value="react">React</Option>
-//   </Select>
-// )
-
-// export const withDescription = (args: any) => (
-//   <Select {...args}>
-//     <Option value="javascript">JavaScript</Option>
-//     <Option value="typeScript">TypeScript</Option>
-//     <Option value="react">React</Option>
-//   </Select>
-// )
-
-// export const size = (args: any) => (
-//   <Select {...args}>
-//     <Option value="javascript">JavaScript</Option>
-//     <Option value="typeScript">TypeScript</Option>
-//     <Option value="react">React</Option>
-//   </Select>
-// )
-
 const data = ['England', 'Wales', 'Scotland', 'Ireland']
 const icon = <IconBook type={'Book'} />
 
-// Default.args = {
-//   disabled: false,
-//   label: 'Label',
-//   className: 'font-sans',
-//   layout: 'vertical',
-//   children: [
-//     <>
-//       <Option value="javascript">JavaScript</Option>
-//       <Option value="typeScript">TypeScript</Option>
-//       <Option value="react">React</Option>
-//     </>,
-//   ],
-// }
-
-// withOptionGroup.args = {
-//   placeholder: 'Type text here ...',
-//   disabled: false,
-//   label: 'Input with an error message',
-//   className: 'font-sans',
-//   value: 'Value of input',
-//   layout: 'vertical',
-// }
-
-// withCheckboxes.args = {
-//   disabled: false,
-//   checkboxes: data,
-//   allowedValues: data,
-//   className: 'font-sans',
-//   layout: 'vertical',
-// }
-
 ErrorState.args = {
-  // placeholder: 'Type text here ...',
-  // disabled: false,
-  // label: 'Input with an error message',
-  // className: 'font-sans',
-  // // value: 'Value of input',
-  // error: 'Your password must be less than 4 characters.',
-  // layout: 'vertical',
-  // defaultValue:people[1].value,
-  // defaultValue: people[1].value,
   label: 'Choose a person',
   descriptionText: 'Choose a person for this role',
 }
-
-// withIcon.args = {
-//   placeholder: 'Type text here ...',
-//   disabled: false,
-//   label: 'Input with an Icon',
-//   className: 'font-sans',
-//   value: 'Value of input',
-//   icon: icon,
-//   allowedValues: data,
-//   layout: 'vertical',
-// }
-
-// withOptionLabel.args = {
-//   placeholder: 'Type text here ...',
-//   disabled: false,
-//   label: 'Input with an error message',
-//   className: 'font-sans',
-//   value: 'Value of input',
-//   labelOptional: 'This is required',
-//   allowedValues: data,
-//   layout: 'vertical',
-// }
-
-// withBeforeAndAfterLabel.args = {
-//   placeholder: 'Type text here ...',
-//   disabled: false,
-//   label: 'Label',
-//   beforeLabel: 'Before : ',
-//   afterLabel: ' : After',
-//   className: 'font-sans',
-//   value: 'Value of input',
-//   labelOptional: 'This is required',
-//   allowedValues: data,
-//   layout: 'vertical',
-// }
-
-// withDescription.args = {
-//   placeholder: 'Type text here ...',
-//   disabled: false,
-//   label: 'Input with an error message',
-//   className: 'font-sans',
-//   value: 'Value of input',
-//   descriptionText: 'Make your password short and easy to guess',
-//   allowedValues: data,
-//   layout: 'vertical',
-// }
-
-// size.args = {
-//   placeholder: 'Type text here ...',
-//   disabled: false,
-//   label: 'Input with a size selected',
-//   value: 'Value of input',
-//   descriptionText: 'Choose a different size and font and padding will change',
-//   allowedValues: data,
-//   layout: 'vertical',
-//   size: 'tiny',
-// }
