+++ src/components/Listbox/index.tsx
@@ -1,0 +1,3 @@
+import Listbox from './Listbox'
+export default Listbox
+export { default as Listbox } from './Listbox'
