--- src/components/SelectStyled/index.tsx
@@ -1,3 +1,0 @@
-import SelectStyled from './SelectStyled'
-export default SelectStyled
-export { default as SelectStyled } from './SelectStyled'
